Backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
opencv-python==4.2.0.34
pandas==0.25.1
pdf2image==1.12.1
Pillow==7.1.2
PyMuPDF==1.17.0
pywin32==227
Werkzeug==0.16.0